move main code into simulator folder
